# entertainmentEmporiumBackend/schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Entertainment Emporium Backend API 
  description: A movies and tv shows rating API 
  contact:
    name: Francesca 
    email: passmoref@coventry.ac.uk
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://hexagonprofile-maximumalex-3000.codio-box.uk/api/v1'
    description: Development server
...
paths:
...
  /movies/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Numeric id of the movie
    get:
      tags:
        - Movies
      description:
        Get movie by id.
      responses:
        '200':
          description: View of a Movie.
        '404':
          description: Movie was not found in the DB.
      
    put:
      tags:
        - Movies
      description:
        Update movie by ID.
      responses:
        '200':
          description: Update confirmation and link to record.
        '404':
          description: Movie was not found in the DB.
        '401':
          description: Admins can only update Movies
        '403':
          description: Requester forbidden to update this resource

    delete:
      tags:
        - Movies
      description:
        Delete movie by ID.
      responses:
        '200':
          description: Delete confirmation.
        '401':
          description: Only admin users can delete movies.
        '404':
          description: Movie was not found in the DB.
        '403':
          description: Requester forbidden to delete this resource
  /movies:
    summary: Represents an individual movie
    description: >
      This resource represents an individual movie in the system.
      Each movie is identified by a numeric `id`.
    post:
      tags:
        - Movies
      description:
        Create a new movie.
      requestBody:
        description: All new movie data.
        required: true
        content:
          application/json:
            schema:
              $ref: ./movies.json#/definitions/movie
      responses:
        '201':
          description: Movie created
          content:
            application/json:
              schema:
                $ref: ./movies.json#/definitions/movieCreated
  /movies/{page}/{limit}/{column}/{order}:
    parameters:
      - in: path
        name: page
        schema:
          type: integer
        required: true
        description: Numeric page of results

      - in: path
        name: limit
        schema:
          type: integer
        required: true
        description: Numeric amount of results to retrieve

      - in: path
        name: column
        schema:
          type: string
        required: true
        description: The column which to order the results 

      - in: path
        name: order
        schema:
          type: string
        required: true
        description: The order which results are displayed 

    get:
      tags:
        - Movies
      description:
        Get multiple movies with pagination.
      responses:
        '200':
          description: View of multiple movies. 
        '404':
          description: Movies were not found in the DB.
  /movies/{title}:
    parameters:
      - in: path
        name: title
        schema:
          type: string
        required: true
        description: Title of the movie 
    get:
      tags:
        - Movies
      description:
        Get movie by the title.
      responses:
        '200':
          description: View of the movie record.
        '404':
          description: Movie was not found in the DB.
...
  /movieReviews/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Numeric id of the movie review
    get:
      tags:
        - Movie Reviews
      description:
        Get movie review by id.
      responses:
        '200':
          description: View of a movie reviews record 
        '404':
          description: Movie review was not found in the DB.
    put:
      tags:
        - Movie Reviews
      description:
        Update movie reviews by ID.
      responses:
        '200':
          description: Update confirmation.
        '404':
          description: Movie review was not found in the DB.
        '401':
          description: Users can only update their own movie reviews
        '403':
          description: Requester forbidden to update this resource

    delete:
      tags:
        - Movie Reviews
      description:
        Delete movie reviews by ID.
      responses:
        '200':
          description: Delete confirmation.
        '401':
          description: Only users can delete their own movie reviews.
        '404':
          description: Movie review was not found in the DB.
        '403':
          description: Requester forbidden to delete this resource
  /movieReviews:
    summary: Represents an individual movie review
    description: >
      This resource represents an individual movie review in the system.
      Each movie review is identified by a numeric `id`.
    post:
      tags:
        - Movie Reviews
      description:
        Create a new movie review account.
      requestBody:
        description: All new movie review data.
        required: true
        content:
          application/json:
            schema:
              $ref: ./movieReviews.json#/definitions/movieReview
      responses:
        '201':
          description: Movie Review created
          content:
            application/json:
              schema:
                $ref: ./movieReviews.json#/definitions/reviewCreated
...